+++ WT-main-project/projectbackend/Routes/UserRoutes.js
@@ -1,38 +1,72 @@
-const express=require('express')
+const express = require('express')
 const user=require('../models/user')
 const userrouter=express.Router();
 
 userrouter.get('/',async (req,res)=>{
-    const data=await user.find();
-    res.send(data)
+    try {
+        const data = await user.find();
+        res.send(data);
+    } catch (error) {
+        res.send(error);
+    }
 })
 
 userrouter.get('/:id',async (req,res)=>{
-    const data=await user.findById(req.params.id);
-    res.send(data)
+    try {
+        const data = await user.findById(req.params.id);
+        if (!data) {
+            return res.send("user not found");
+        }
+        res.send(data);
+    } catch (error) {
+        return res.send(error);
+    }
 })
 
 userrouter.post('/', async (req, res) => {
     if (!req.body || Object.keys(req.body).length === 0) {
-        return res.status(400).send("Please provide the required data");
+        return res.send("Please provide the required data");
     }
     try {
         const data = await user.create(req.body);
-       res.send(data)
+        res.send(data)
     } catch (error) {
-        return res.send(error)
-        // return res.status(500).send({ error: "Error creating user", details: error.message });
+        return res.send(error);
     }
 });
 
 
 userrouter.patch('/:id',async (req,res)=>{
-    const data=await user.findByIdAndUpdate(req.params.id,req.body);
-    res.send(data)
+    if (!req.body || Object.keys(req.body).length === 0) {
+        return res.send("Please provide data to update");
+    }
+    try {
+        const data = await user.findByIdAndUpdate(req.params.id, req.body, { new: true });
+        if (!data) {
+            return res.send("user not found");
+        }
+        res.send(data);
+    } catch (error) {
+        return res.send(error);
+    }
 })
 
-userrouter.delete('/:id',async (req,res)=>{
-    const data=await user.findByIdAndDelete(req.params.id);
-    res.send(data)
+userrouter.delete('/:id?',async (req,res)=>{
+    
+    console.log(req.params.id);
+    if (!req.params.id) {
+        return res.send("Please provide an ID");
+    }
+    
+    try {
+        const data = await user.findByIdAndDelete(req.params.id);
+        if (!data) {
+            return res.send("user not found");
+        }
+        res.send(data);
+    } catch (error) {
+        return res.send(error);
+    }
+    
 })
 module.exports=userrouter;
